Move area components to Area folder
