+++ packages/x-charts/src/LineChart/Area/AreaElement.tsx
@@ -4,11 +4,11 @@ import composeClasses from '@mui/utils/composeClasses';
 import useSlotProps from '@mui/utils/useSlotProps';
 import generateUtilityClass from '@mui/utils/generateUtilityClass';
 import generateUtilityClasses from '@mui/utils/generateUtilityClasses';
-import { SlotComponentPropsFromProps } from '../internals/SlotComponentPropsFromProps';
-import { useInteractionItemProps } from '../hooks/useInteractionItemProps';
+import { SlotComponentPropsFromProps } from '../../internals/SlotComponentPropsFromProps';
+import { useInteractionItemProps } from '../../hooks/useInteractionItemProps';
 import { AnimatedArea, AnimatedAreaProps } from './AnimatedArea';
-import { SeriesId } from '../models/seriesType/common';
-import { useItemHighlighted } from '../context';
+import { SeriesId } from '../../models/seriesType/common';
+import { useItemHighlighted } from '../../context/HighlightedProvider';
 
 export interface AreaElementClasses {
   /** Styles applied to the root element. */
