+++ packages/x-charts/src/LineChart/Area/index.ts
@@ -1,0 +1,3 @@
+export * from './AnimatedArea';
+export * from './AreaElement';
+export * from './AreaPlot';
