+++ packages/x-charts/src/LineChart/index.ts
@@ -1,12 +1,10 @@
 export * from './LineChart';
 
 export * from './LinePlot';
-export * from './AreaPlot';
+export * from './Area';
 export * from './MarkPlot';
 export * from './LineHighlightPlot';
 
-export * from './AreaElement';
-export * from './AnimatedArea';
 export * from './LineElement';
 export * from './AnimatedLine';
 export * from './MarkElement';
